UB-1405: unmount flow should not fail if discovery fails (#229)
In this PR , the following idempotent issue with Unmount flow is fixed:
if a discovery of a volume fails, no need to fail the Unmount command, just need to skip the UnmountDeviceFlow.

more info:
if volume's mpath NOT discovered (because the device does not exist at all in multipath -ll) then we should skip UnmountDeviceFlow instead of failing (reminder - the UnmountDeviceFlow is the flow that does umount the device and cleanup of the mpath device) .
Why its idempotent? - because there is a chance that unmount flow will fail at the end of UnmountDeviceFlow after it unmounts and cleans the mpath device. So to be idempotent the unmount needs to handle unmount request while the mpath device was already cleaned.