UB-1406 remove mountpoint safely (#248)
in this PR we want to check before deleting the /ubiquity/wwn folder if there are any files left there. if there are still files(e.g for some reason the mountpoint directory /ubiquity/wwn had some files before the block device mountpoint on it) - an ERROR will be raised until the folder will be cleaned of files.